refactor: 인증 타입 시스템 통합 및 중복 제거 - Auth: 타입 정의를 auth.types.ts로 통합 - Refactor: 관련된 모든 컨트롤러, 서비스, 전략(Strategy) 파일의 타입 참조 수정
