+++ src/auth/strategies/jwt.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable, UnauthorizedException } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
 import { PassportStrategy } from "@nestjs/passport";
 import { ExtractJwt, Strategy } from "passport-jwt";
-import { GuestJwtPayload, UserJwtPayload } from "../types/auth.types";
+import { AuthenticatedUser, GuestJwtPayload, GuestUser, UserJwtPayload } from "../types/auth.types";
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, "jwt") {
@@ -14,7 +14,7 @@ export class JwtStrategy extends PassportStrategy(Strategy, "jwt") {
     });
   }
 
-  validate(payload: UserJwtPayload | GuestJwtPayload) {
+  validate(payload: UserJwtPayload | GuestJwtPayload): AuthenticatedUser | GuestUser {
     if ("isGuest" in payload) {
       return {
         ip: payload.ip,
