+++ src/factcheck/factcheck.controller.ts
@@ -11,6 +11,7 @@ import {
   Query,
   Request,
 } from "@nestjs/common";
+import type { AuthenticatedUser, RequestWithUser } from "../auth/types/auth.types";
 import { CurrentUser } from "../common/decorators/current-user.decorator";
 import { RequireLogin } from "../common/decorators/require-login.decorator";
 import { CreateFactCheckDto } from "./dto/create-factcheck.dto";
@@ -18,7 +19,6 @@ import type { FactCheckListResponse } from "./dto/factcheck-list-response.dto";
 import type { ClaimStatusUpdateResponse, FactCheckResponse } from "./dto/factcheck-response.dto";
 import { GetFactCheckListQueryDto } from "./dto/pagination-query.dto";
 import { FactCheckService } from "./factcheck.service";
-import type { AuthenticatedUser, RequestWithUser } from "./types/factcheck.types";
 
 @Controller("factcheck")
 export class FactCheckController {
@@ -39,7 +39,7 @@ export class FactCheckController {
     @CurrentUser() user: AuthenticatedUser,
     @Query() query: GetFactCheckListQueryDto,
   ): Promise<FactCheckListResponse> {
-    return this.factCheckService.getFactCheckList(user, query);
+    return this.factCheckService.getFactCheckHistory(user, query);
   }
 
   @Get(":id")
