--- src/factcheck/types/factcheck.types.ts
@@ -1,16 +1,0 @@
-export interface AuthenticatedUser {
-  userId: string;
-  email: string;
-  isGuest: false;
-}
-
-export interface GuestUser {
-  ip: string;
-  isGuest: true;
-}
-
-export type RequestUser = AuthenticatedUser | GuestUser;
-
-export interface RequestWithUser {
-  user: RequestUser;
-}
